fix: resolve build errors across monorepo (#544)
fix: resolve build errors across bot, design, vaults, and sdk

- vaults: add createMDX() wrapper for Turbopack MDX support
- design: switch /api/slides/generate from edge to nodejs runtime
- bot: extract registerApproval from route to lib/supervisor.ts
- bot: fix ToolSet type and inputSchema union inference
- bot: move providerOptions from generate() to ToolLoopAgent constructor
- bot: import AgentSandbox from @syner/vercel instead of @vercel/sandbox
- sdk: refactor getModel/getModelFallbacks into resolveModel(tier)
