+++ apps/vaults/next.config.ts
@@ -1,5 +1,8 @@
+import { createMDX } from "fumadocs-mdx/next";
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {};
 
-export default nextConfig;
+const withMDX = createMDX();
+
+export default withMDX(nextConfig);
