+++ packages/sdk/src/agents/index.ts
@@ -1,3 +1,4 @@
 export { getAgentsRegistry, getAgentsList, getAgentByName, getAgentsByChannel, invalidateAgentsCache } from './loader'
 export type { AgentCard } from './loader'
-export { getModel, getModelFallbacks, MODEL_IDS, FALLBACK_MODELS } from './models'
+export { resolveModel, MODEL_IDS, FALLBACK_MODELS } from './models'
+export type { ModelTier } from './models'
